add failing test when explicitly set layout to false in _data.json
When explicitly set layout to false in _data.json and put a file
in the outside directory with the same name as the directory name
and a template extension, that file will be used as layout. In the
test case, the file is false.jade outside the false/ directory.

This is because false is not explicitly checked in the layout
lookup, and is treated as empty string.