Refactors helpers/raw.js::outputPath to allow dot characters on file name and path.
With the old implementation, outputPath would judge the type of extension by splitting the string with the "dot" character.
This would cause wrong behaviour if the file path contained dots not in the extension, e.g. "foobar-1.0.0.js" would output "foobar-1.0.0", with no extension.
I have also created the option to disallow alternate extensions, which should be the default in my opinion. However, I have left it allowed to maintain compatibility.